Fixed how card info was added to inventory
