+++ client/src/styles.css
@@ -91,7 +91,12 @@ html {
 }
 
 .panel-slider {
-  height: 100vh;
   width: 280px;
   float: left;
+  position: fixed;
+  top: 0%;
+}
+
+.navbar {
+  z-index: 10;
 }
