+++ server/lib/check-if-card-exists.ts
@@ -1,0 +1,13 @@
+import CardRepository from '../repositories/card-repository';
+
+const checkIfCardExists = async (cardNumber: string): Promise<boolean> => {
+  const cardRepository = new CardRepository();
+
+  const card = await cardRepository.getCard(cardNumber);
+
+  console.log(card);
+
+  return false;
+};
+
+export { checkIfCardExists };
